Fix connect to peer button
