+++ packages/browser-client/src/Components/Footer.tsx
@@ -1,20 +1,8 @@
-import { Text, HStack, IconButton, Link, VStack, Button, Input } from '@chakra-ui/react'
-import React, { useContext, useState } from 'react'
+import { Text, HStack, IconButton, Link } from '@chakra-ui/react'
+import React from 'react'
 import { FaDiscord, FaGithub, FaGithubSquare, FaTwitter } from 'react-icons/fa'
-import { AppContext, AppContextType, StateChange } from '../globalReducer'
 
 export default function Footer() {
-  const { state, dispatch } = useContext(AppContext as React.Context<AppContextType>)
-  const [enr, setEnr] = useState('')
-
-  async function connectToPeer() {
-    try {
-      await state.provider?.historyProtocol.addBootNode(state.searchEnr)
-      dispatch({ type: StateChange.SETSEARCHENR, payload: '' })
-      dispatch({ type: StateChange.REFRESHPEERS })
-    } catch (err) {}
-  }
-
   return (
     <HStack marginY={1} width={'100%'} bg={'gray.100'} justifyContent={'space-evenly'}>
       <Link href="https://twitter.com/EthereumJs">
